Refactor: Enhance user caching and role serialization
The `UserService` has been refactored to use a private `_updateCache` method, improving how users are added and updated in the cache. A `logout` method was also introduced to remove a user's profile from the cache upon logging out.

In the `Profile` model, the serialization of the `role` field has been updated to use the `.name` property and defaults to `UserRole.user` if the role is null.