Feat: Add facility filter to room search
A `facilityIds` filter has been added to the `getAll` and `getAvailable` methods in the `RoomService`. This allows filtering rooms to only include those that have all of the specified facility IDs. The filtering logic uses an "AND" condition, meaning a room must contain all selected facilities to be included in the results.